Changed way app was rendering pokemon inside of the card
I am not actually a fan of this way and preferred the way I had it before. It seems super contrived to have the entire app know about every pokemon, rather than just having the card itself dictate whether those pokemon were already there and whether the fetch call needed to happen again. Willing to talk about it more to understand why this way is better, but it just seems more convoluted and needlessly so.